requirements.txt
alabaster==0.7.13
appdirs==1.4.4
attrs==23.1.0
Babel==2.13.1
black==18.9b0
certifi==2023.7.22
charset-normalizer==3.3.2
click==8.1.7
docutils==0.20.1
idna==3.4
imagesize==1.4.1
importlib-metadata==6.7.0
Jinja2==2.11.3
MarkupSafe==2.1.3
more-itertools==9.1.0
packaging==23.2
parsimonious==0.7.0
pluggy==0.13.1
py==1.11.0
Pygments==2.16.1
pytest==5.3.5
python-dateutil==2.8.2
pytz==2023.3.post1
requests==2.31.0
six==1.16.0
snowballstemmer==2.2.0
Sphinx==1.8.3
sphinx-js==2.7.1
sphinxcontrib-serializinghtml==1.1.5
sphinxcontrib-websupport==1.2.4
toml==0.10.2
types-python-dateutil==2.8.19.14
typing_extensions==4.7.1
urllib3==2.0.7
wcwidth==0.2.9
zipp==3.15.0
